pyproject.toml
[project]
name = "pratice-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "completion>=0.1.2",
    "litellm>=1.73.6",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
